voice-agent/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
livekit-agents==1.2.15
livekit-plugins-openai==1.2.15
livekit-plugins-deepgram==1.2.15
livekit-plugins-groq==1.2.15
livekit-plugins-silero==1.2.15
livekit-plugins-turn-detector==1.2.15
python-dotenv==1.0.1
torch==2.4.1+cpu